feat: make cookie SameSite attribute configurable
